+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/TrackedFeaturesTest.java
@@ -43,6 +43,7 @@ public class TrackedFeaturesTest {
             .put(TrackedFeatures.ANALYTICS_CORE_ENABLED, 13)
             .put(TrackedFeatures.FAST_FAIL_ON_NOT_FOUND_ENABLED, 14)
             .put(TrackedFeatures.MULTI_OPEN_API, 15)
+            .put(TrackedFeatures.COMPOSE_DELETE_SOURCE_ENABLED, 16)
             .build();
 
     // Verifies that no new features have been added without updating this test.