+++ gcs/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Next
 1. Skip soft delete for compose operations using deleteSourceObjects
+1. Optimize integration tests by enabling shared bucket reuse
 
 ## 3.1.19 - 2026-09-22
 1. Add fs.gs.universe.domain support for Google multi-universe (TPC)