+++ gcs/CHANGES.md
@@ -1,5 +1,7 @@
 # Release Notes
 
+## Next
+
 ## 4.0.6 - 2026-09-24
 1. Skip soft delete for compose operations using deleteSourceObjects
 1. Upgrade google.api-client-libraries.version to 2.9.0